thuc_hanh/pyproject.toml
[project]
name = "thuc-hanh"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pycryptodome>=3.23.0",
    "starhtml[debug]>=0.6.0",
]

[tool.ruff.lint]
select = ["F401"]
ignore = ["F403", "F405"]

[tool.ty.rules]
unresolved-reference = "ignore"
